1. Fix the minimizer mask issue; 2. Change the mapping quality formula; 3. Lower minClustersize to map more reads.

+++ LogLookUpTable.h
@@ -7,8 +7,7 @@
 void 
 CreateLookUpTable(std::vector<float> & LookUpTable){
 	for (int i = 501; i <= 10001; i = i + 5) {
-		//cerr << "i: " << i << endl;
-		LookUpTable.push_back(std::log(i));
+		LookUpTable.push_back(logf(i));
 	}
 
 }
